+++ src/main/java/ni/org/ics/estudios/domain/Pbmc/Pbmc.java
@@ -22,7 +22,7 @@ public class Pbmc extends BaseMetaData implements Serializable{
     private Character pbmc_tiene_serologia ='0';
     private Character enviado ='0';
     private String estudios;
-    private Integer edadMeses;
+    private double edadMeses;
     private String notas;
     private String casaCHF;
     private Integer casaPDCS;
@@ -105,11 +105,11 @@ public class Pbmc extends BaseMetaData implements Serializable{
     }
 
     @Column(name = "EDAD_MESES", nullable = true)
-    public Integer getEdadMeses() {
+    public double getEdadMeses() {
         return edadMeses;
     }
 
-    public void setEdadMeses(Integer edadMeses) {
+    public void setEdadMeses(double edadMeses) {
         this.edadMeses = edadMeses;
     }
